refactor(reply): remove retired follow-up exports
