+++ src/auto-reply/reply/followup-delivery.ts
@@ -1,6 +1,5 @@
 /** Prepares queued follow-up payloads for source-channel delivery. */
 import { hasOutboundReplyContent } from "openclaw/plugin-sdk/reply-payload";
-import type { MessagingToolSend } from "../../agents/embedded-agent-messaging.types.js";
 import {
   hasCommittedSourceReplyDeliveryEvidence,
   hasCompletedSourceReplyDeliveryEvidence,
@@ -9,42 +8,24 @@ import {
 } from "../../agents/embedded-agent-runner/delivery-evidence.js";
 import { hasDeliberateSilentTerminalReply } from "../../agents/embedded-agent-runner/result-fallback-classifier.js";
 import { buildAgentRuntimeDeliveryPlan } from "../../agents/runtime-plan/build.js";
-import type { ReplyToMode } from "../../config/types.base.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { logVerbose } from "../../globals.js";
 import { defaultRuntime } from "../../runtime.js";
 import { sessionDeliveryChannel } from "../../utils/delivery-context.shared.js";
-import { stripHeartbeatToken } from "../heartbeat.js";
-import {
-  copyReplyPayloadMetadata,
-  getReplyPayloadMetadata,
-  isReplyPayloadStatusNotice,
-  setReplyPayloadMetadata,
-} from "../reply-payload.js";
-import type { OriginatingChannelType } from "../templating.js";
+import { getReplyPayloadMetadata, isReplyPayloadStatusNotice } from "../reply-payload.js";
 import type { ReplyPayload } from "../types.js";
 import { normalizeAssistantFinalDeliveryText } from "./agent-runner-core.js";
 import type { AgentTurnExecutionResult } from "./agent-runner-execution.types.js";
 import { buildEmptyInteractiveReplyPayload } from "./agent-runner-failure-reply.js";
 import type { AccountedAgentTurn } from "./agent-runner-result-accounting.js";
 import { appendUsageLine, resolveResponseUsageLine } from "./agent-runner-usage-line.js";
+import { resolveFollowupDeliveryPayloads } from "./followup-delivery-payloads.js";
 import type { AdmittedFollowupTurn, FollowupRunnerParams } from "./followup-turn-admission.js";
 import type { InternalGetReplyOptions } from "./get-reply.types.js";
-import {
-  resolveOriginAccountId,
-  resolveOriginMessageProvider,
-  resolveOriginMessageTo,
-} from "./origin-routing.js";
+import { resolveOriginMessageProvider } from "./origin-routing.js";
 import { warnPrivateMessageToolFinal } from "./private-message-tool-final.js";
 import { enqueueFollowupRun, resolveQueueSettings, type FollowupRun } from "./queue.js";
 import type { ReplyDispatchKind } from "./reply-dispatcher.types.js";
-import {
-  applyReplyThreading,
-  filterMessagingToolDuplicates,
-  filterMessagingToolMediaDuplicates,
-  resolveMessagingToolPayloadDedupe,
-} from "./reply-payloads.js";
-import { createReplyDeliveryContext, resolveReplyToMode } from "./reply-threading.js";
 import { isRoutableChannel, routeReply } from "./route-reply.js";
 import { resolveSourceReplyVisibilityPolicy } from "./source-reply-delivery-mode.js";
 import {
@@ -52,124 +33,6 @@ import {
   resolveStrandedReplyRecovery,
 } from "./stranded-reply-recovery.js";
 import { createTypingSignaler } from "./typing-mode.js";
-
-/** Strips empty/heartbeat payloads, applies threading, and dedupes message-tool sends. */
-export function resolveFollowupDeliveryPayloads(params: {
-  cfg: OpenClawConfig;
-  payloads: ReplyPayload[];
-  messageProvider?: string;
-  originatingAccountId?: string;
-  originatingChannel?: string;
-  originatingChatType?: string | null;
-  originatingReplyToMode?: ReplyToMode;
-  originatingTo?: string;
-  originatingThreadId?: string | number;
-  reasoningPayloadsEnabled?: boolean;
-  commentaryPayloadsEnabled?: boolean;
-  sentMediaUrls?: string[];
-  sentTargets?: MessagingToolSend[];
-  sentTexts?: string[];
-}): ReplyPayload[] {
-  const replyMessageProvider = resolveOriginMessageProvider({
-    originatingChannel: params.originatingChannel,
-    provider: params.messageProvider,
-  });
-  const replyToChannel = replyMessageProvider as OriginatingChannelType | undefined;
-  const replyToMode =
-    params.originatingReplyToMode ??
-    resolveReplyToMode(
-      params.cfg,
-      replyToChannel,
-      params.originatingAccountId,
-      params.originatingChatType,
-    );
-  const accountId = resolveOriginAccountId({
-    originatingAccountId: params.originatingAccountId,
-  });
-  const replyDelivery = createReplyDeliveryContext(replyToMode, params.originatingChatType);
-  const replyDeliverySource = replyMessageProvider
-    ? {
-        channel: replyMessageProvider,
-        ...(accountId ? { accountId } : {}),
-      }
-    : undefined;
-  const deliverablePayloads = params.payloads.filter(
-    (payload) =>
-      !(payload.isReasoning === true && params.reasoningPayloadsEnabled !== true) &&
-      !(payload.isCommentary === true && params.commentaryPayloadsEnabled !== true),
-  );
-  const sanitizedPayloads: ReplyPayload[] = [];
-  for (const payload of deliverablePayloads) {
-    const text = payload.text;
-    const sanitized =
-      text?.includes("HEARTBEAT_OK") === true
-        ? copyReplyPayloadMetadata(payload, {
-            ...payload,
-            text: stripHeartbeatToken(text, { mode: "message" }).text,
-          })
-        : payload;
-    // Normalize before callers decide whether the run was empty. Otherwise a
-    // whitespace-only model payload can suppress the interactive fallback.
-    if (hasOutboundReplyContent(sanitized, { trimText: true })) {
-      sanitizedPayloads.push(sanitized);
-    }
-  }
-  const replyTaggedPayloads = applyReplyThreading({
-    payloads: sanitizedPayloads,
-    replyToMode,
-    replyToChannel,
-  }).map((payload) =>
-    setReplyPayloadMetadata(payload, {
-      replyDelivery,
-      ...(replyDeliverySource ? { replyDeliverySource } : {}),
-    }),
-  );
-  const sentMediaUrlFallback = params.sentMediaUrls ?? [];
-  const sentTextFallback = params.sentTexts ?? [];
-  const originatingTo = resolveOriginMessageTo({
-    originatingTo: params.originatingTo,
-  });
-  const dedupedPayloads: ReplyPayload[] = [];
-  for (const payload of replyTaggedPayloads) {
-    const decision = resolveMessagingToolPayloadDedupe({
-      config: params.cfg,
-      messageProvider: replyMessageProvider,
-      messagingToolSentTargets: params.sentTargets,
-      originatingTo,
-      originatingThreadId: params.originatingThreadId,
-      replyToId: payload.replyToId,
-      replyToIsExplicit: Boolean(
-        getReplyPayloadMetadata(payload)?.replyToIdExplicit ||
-        payload.replyToTag ||
-        payload.replyToCurrent,
-      ),
-      replyDelivery: getReplyPayloadMetadata(payload)?.replyDelivery,
-      accountId,
-    });
-    if (!decision.shouldDedupePayloads) {
-      dedupedPayloads.push(payload);
-      continue;
-    }
-    const sentMediaUrls =
-      decision.matchingRoute && !decision.useGlobalSentMediaUrlEvidenceFallback
-        ? decision.routeSentMediaUrls
-        : sentMediaUrlFallback;
-    const sentTexts =
-      decision.matchingRoute && !decision.useGlobalSentTextEvidenceFallback
-        ? decision.routeSentTexts
-        : sentTextFallback;
-    const mediaFiltered = filterMessagingToolMediaDuplicates({
-      payloads: [payload],
-      sentMediaUrls,
-    });
-    const textFiltered = filterMessagingToolDuplicates({
-      payloads: mediaFiltered,
-      sentTexts,
-    });
-    dedupedPayloads.push(...textFiltered);
-  }
-  return dedupedPayloads;
-}
 
 type FollowupDeliveryDecision =
   | {
